requirements.txt
livekit-agents>=1.0.0
livekit-plugins-anthropic>=1.0.0
livekit-plugins-deepgram>=1.0.0
livekit-plugins-elevenlabs>=1.0.0
livekit-plugins-silero>=1.0.0
python-dotenv>=1.0.0
anthropic>=0.40.0
gspread>=6.0.0
google-auth>=2.0.0
twilio>=9.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
python-multipart>=0.0.7
